+++ Controllers/API/OrderReport/KBNRT190Controller.cs
@@ -79,6 +79,7 @@ namespace KANBAN.Controllers.API.OrderReport
         {
             try
             {
+                setConString();
                 string _result = "";
                 var tripDB = await _KB3Context.V_KBNRT_190_Rpt.Select(x => x.Deli_trip).Distinct().OrderBy(x => x).ToListAsync();
 
@@ -108,6 +109,7 @@ namespace KANBAN.Controllers.API.OrderReport
             }
             try
             {
+                setConString();
                 string _result = "";
                 dynamic _json = JsonConvert.DeserializeObject<dynamic>(data);
                 string dateFrom = _json["dateFrom"];
@@ -147,6 +149,7 @@ namespace KANBAN.Controllers.API.OrderReport
             }
             try
             {
+                setConString();
                 string _result = "";
                 dynamic _json = JsonConvert.DeserializeObject(data);
                 string supFrom = _json["supFrom"];
@@ -165,6 +168,11 @@ namespace KANBAN.Controllers.API.OrderReport
                 string UserName = HttpContext.Session.GetString("USER_NAME");
                 string HostName = HttpContext.Session.GetString("USER_DEVICENAME");
 
+                if (string.IsNullOrWhiteSpace(UserName) || string.IsNullOrWhiteSpace(HostName))
+                {
+                    return Redirect($"{Request.Path.ToString()}{Request.QueryString.Value.ToString()}");
+                }
+
                 await _KB3Context.Database.ExecuteSqlRawAsync("DELETE FROM RPT_KBNRT_190 WHERE F_Update_By = @UserName " +
                     " AND F_Host_name = @HostName",
                     new SqlParameter("@UserName", UserName),
@@ -179,6 +187,18 @@ namespace KANBAN.Controllers.API.OrderReport
                     .Where(x => x.Deli_trip >= tripFrom && x.Deli_trip <= tripTo)
                     .ToListAsync();
 
+                if(rptList.Count == 0)
+                {
+                    _result = @"{
+                                    ""status"":""404"",
+                                    ""response"":""OK"",
+                                    ""title"":""Report Data Not Found"",
+                                    ""message"": ""Please Try Other Option!""
+                                    }";
+
+                    return Ok(_result);
+                }
+
                 foreach (var rpt in rptList)
                 {
                     await _KB3Context.Database.ExecuteSqlRawAsync("INSERT INTO RPT_KBNRT_190 (Supplier_cd, Store_cd, Status, PDS_no, KB_no, " +
@@ -188,6 +208,8 @@ namespace KANBAN.Controllers.API.OrderReport
                         rpt.Deli_Date, rpt.Deli_trip, rpt.F_Box_Qty, rpt.Qty_KB, rpt.F_Unit_Amount, rpt.F_Receive_amount, UserName, HostName
                         );
                 }
+
+
 
                 string _jsonData = JsonConvert.SerializeObject(UserName);
                 string _jsonData2 = JsonConvert.SerializeObject(HostName);
@@ -216,6 +238,7 @@ namespace KANBAN.Controllers.API.OrderReport
             }
             try
             {
+                setConString();
                 string _result = "";
                 dynamic _json = JsonConvert.DeserializeObject(data);
                 string supFrom = _json["supFrom"];
@@ -244,6 +267,18 @@ namespace KANBAN.Controllers.API.OrderReport
                     .Where(x => x.Prt_no.CompareTo(partFrom) >= 0 && x.Prt_no.CompareTo(partTo) <= 0)
                     .Where(x => x.chk_Order_Date.CompareTo(dateFrom) >= 0 && x.chk_Order_Date.CompareTo(dateTo) <= 0)
                     .ToListAsync();
+
+                if (rptList.Count == 0)
+                {
+                    _result = @"{
+                                    ""status"":""404"",
+                                    ""response"":""OK"",
+                                    ""title"":""Report Data Not Found"",
+                                    ""message"": ""Please try other option!""
+                                    }";
+
+                    return Ok(_result);
+                }
 
                 foreach (var rpt in rptList)
                 {
